Added a custom __deepcopy__ method to the RammbockCore class
The custom method does not copy the _clients and _servers attributes,
which are local caches and may contain objects of type 'socket'

This is a first shot at fixing the load_copy_of_template() method,
which failed because of the 'socket' objects.

The final solution would probably be to inhibit (deep)copying the
RammbockCore class altogether.